The scale of variance depends on the mean for both sigma^2 and lambda_0. Not true for position.
Our method handles appearance/disappearance of cluster.
Thinning and fattening of cluster also.  All 3 parameters are now allowed change.

The only thing that doesn't work is when we have N clusters but not enough data, and ground truth clusters are fit as 1 merged one.  It may be that the remaining cluster has been set to have a very large variance along the way, and can't be brought back to normal.  This might be something to check for in the future.