Working now.  q2 and Sg are updated differently.
q2 is updated even if nSpksM == 0, because NO spikes is informative about the PF.  However, it is not informative about the marks.

If I see no spikes, and I have information about where I've been, that provides information about PF shape.
However, this tells me nothing about its shape in mark space.

There q2 can be updated absent spike being assigned to this cluster, Sg cannot.

Because of this, if I have a cluster that is not assigned to, spatially it becomes very wide, while the uncertainty remains small.

So we are going to manually check if cluster has small FR or its center is way out of range.  If so, we reset it to have large uncertainty.

Now, when going to the next ITER, we check new data to see how far they are from model.  If we find a point that largely deviates,  we will move
the IC of the uncertain cluster near these points.  That is next checkin.