fitutil does sepHash better.
Also, working gibbsApprMXM.py      works for 15 clusters with new place fields appearing mid-experiment in 2D marks.  Clusters that have no marks
assigned to them are reset to be fairly narrow centers, but with high uncertainty about where the center mean is.

Want to now set the initial condition of those uncertain centers to be near marks that are far from current significant centers.