Broken gibbsApprMxM.py for real data probably fixed.  Using the inverse-gamma distribution with alpha < 1 gives us non-normalizable distribution, which together with with silence term that favors big variances, blurs out clusters.  Use IG distribution with alpha > 1 as priors.
The spatial mean prior is 0 with a large variance
The mark mean prior is mean of marks with a large covariance  (can be anywhere)

The spatial variance prior.  We know something about the width of place fields.  Unlike mean of mean, spatial width has a characteristic size.  We know they shouldn't be too large - meaning we should at the minimum choose a prior which has a mean.
The mark variance prior.  At most, mark cluster is as wide as the unclusterd marks.  Ditto here.