Major changes:
1)  sample from CDF for q2.  Instead of trying to read off a, B from conditional posterior, just sample from the CDF, which isn't inverse gamma.
At the end, look at the samples.  That posterior should be roughly inverse gamma.  Read the hyper-parameters from the posterior distribution.

2)  invert order of sampling.  l0 before q2.  q2 is sampled numerically, and has hard limits on its range.  l0 and q2 are correlated.  If l0 goes off scale, this can cause
q2 to be truncatedly sampled.  If we reverse order, no matter what value q2 takes, l0 because it is analytic, can be sampled appropriately.  Hopefully this will make the sampler
numerically stable.